Move task templating into its own module
