+++ crates/project/src/task_inventory.rs
@@ -10,7 +10,7 @@ use collections::{HashMap, VecDeque};
 use gpui::{AppContext, Context, Model, ModelContext, Subscription};
 use itertools::Itertools;
 use language::Language;
-use task::{ResolvedTask, Task, TaskId, TaskSource, TaskTemplate};
+use task::{from_template_definition, ResolvedTask, Task, TaskId, TaskSource};
 use util::{post_inc, NumericPrefixWithSuffix};
 use worktree::WorktreeId;
 
@@ -145,14 +145,13 @@ impl Inventory {
             .into_iter()
             .flat_map(|(tasks, language)| {
                 tasks.0.into_iter().map(move |definition| {
-                    let language_name = language.name();
-                    TaskTemplate {
-                        id: TaskId(format!("language_{language_name}")),
+                    from_template_definition(
+                        TaskId(format!("language_{}", language.name())),
                         definition,
-                    }
+                    )
                 })
             })
-            .flat_map(|task| Some((task_source_kind.as_ref()?, Arc::new(task) as Arc<dyn Task>)));
+            .flat_map(|task| Some((task_source_kind.as_ref()?, task as Arc<dyn Task>)));
 
         let mut lru_score = 0_u32;
         let tasks_by_usage = if lru {
